Don't auto merge default branch into the deploying branch
More info what `auto_merge` attribute does:
https://docs.github.com/en/rest/deployments/deployments?apiVersion=2022-11-28#create-a-deployment--parameters

When we deploy from the default repo branch this change won't make any difference, there is no need to auto merge the default branch into the branch being deployed since they are the same.

However, if we deploy from a non-default branch - like `the-end-is-near` (that we use for feature freezing legacy apps), then we need to ensure that the default branch (`main`) doesn't get auto merged into the deploying branch (`the-end-is-near`).