Pass secrets via `secrets:`
As far as I understand, if you pass ${{ secrets.foo }} to with:, GitHub
will mask it because it understands it is a secret. I guess it wont be
masked if you pass something that should be secret, but it is not coming
from the secrets store.

Also, my understanding is that, this will make us able to do

    secrets: inherit

in calling workflows, if we don't want to pass each secret explicitly
(would work for our org secrets as they are named the same in all
repos).

Relevant docs:

- https://docs.github.com/en/actions/how-tos/reuse-automations/reuse-workflows#using-inputs-and-secrets-in-a-reusable-workflow
- https://docs.github.com/en/actions/reference/workflows-and-actions/workflow-syntax#onworkflow_callsecrets
- https://docs.github.com/en/actions/reference/security/secure-use